refactor and add proptypes
